add views for mvc mode
