+++ app/templates/base.html
@@ -17,7 +17,7 @@
 <nav class="navbar navbar-default">
     <div class="container-fluid">
         {% if not session.logged_in %}
-        <a href="{{ url_for('login') }}">log in</a>
+        <a href="{{ url_for('login.index') }}">log in</a>
         {% else %}
         <a href="{{ url_for('logout') }}">log out</a>
         {% endif %}
