+++ manage.py
@@ -1,23 +1,21 @@
 #!/usr/bin/env python
 import os
-from datetime import datetime
 
-from flask import abort
-from flask import flash
 from flask import make_response, jsonify
 from flask import render_template
-from flask import request, session, redirect, url_for
 from flask_migrate import Migrate, MigrateCommand
 from flask_script import Manager, Shell
 
 from app import create_app, db
-from app.forms import LoginForm
 from app.model import Todo
 
 app = create_app(os.getenv('FLASK_CONFIG') or 'default')
 
 manager = Manager(app)
 migrate = Migrate(app, db)
+app_dir = os.path.abspath(os.path.join(os.path.dirname(__file__), 'app'))
+
+print(app_dir)
 
 
 def make_shell_context():
@@ -37,62 +35,8 @@ def test():
 
 
 @app.route('/test', methods=['GET'])
-def test():
+def test_html():
     return render_template('test.html'), 200
-
-
-@app.route('/login', methods=['GET', 'POST'])
-def login():
-    form = LoginForm()
-    if form.validate_on_submit():
-
-        if form.username.data != 'user' or form.password.data != 'pass':
-            flash('Invalid username or password, please try again.')
-        else:
-            flash('login successfully')
-            session['logged_in'] = True
-            return redirect(url_for('todos'))
-    return render_template('login.html', form=form)
-
-
-@app.route('/add', methods=['POST'])
-def add_item():
-    if not session['logged_in']:
-        abort(401)
-
-    td = Todo(request.form['title'])
-    db.session.add(td)
-    db.session.commit()
-
-    flash('new entry added successfully')
-    return redirect(url_for('todos'))
-
-
-@app.route('/del/<int:id>', methods=['POST'])
-def del_item(id):
-    print("delete item %d" % id)
-
-    td = Todo.query.filter_by(id=id).first()
-    db.session.delete(td)
-    db.session.commit()
-
-    return redirect(url_for('todos'))
-
-
-@app.route('/')
-def todos():
-    if not session.get('logged_in'):
-        return redirect(url_for('login'))
-
-    items = Todo.query.all()
-    return render_template('todo_list.html', entries=items, current_time=datetime.utcnow())
-
-
-@app.route('/logout')
-def logout():
-    session.pop('logged_in', None)
-    flash('You were logged out')
-    return redirect(url_for('login'))
 
 
 @app.errorhandler(404)
@@ -106,7 +50,7 @@ def get_all_todo():
 
     json_array = []
     for i in items:
-        json_array.append({"item" : i.title})
+        json_array.append({"item": i.title})
 
     return jsonify({"items": json_array})
